It was decided to move plotting into separate files. It takes 600ms, potentially problem for Xana
